[EDIT] Refactored AnodeType from enum to class with static strings

+++ Core/Entities/Packets/Models/DTO/Announcements/DTOAnnouncementMethod.cs
@@ -7,6 +7,6 @@ public partial class DTOAnnouncement : DTOPacket, IDTO<Announcement, DTOAnnounce
 {
 	public DTOAnnouncement(Announcement announcement) : base(announcement)
 	{
-		AnodeType = (int)announcement.AnodeType;
+		AnodeType = announcement.AnodeType;
 	}
 }
